requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
websockets
ibm-watson
ibm-cloud-sdk-core
groq
twilio
